+++ PhotoManager/PhotoManager.Constants/AssetConstants.cs
@@ -6,7 +6,7 @@ public static class AssetConstants
     // Performances are decreased with PHash by 6 times (for ex: 0.17s for 140 pictures with SHA512 and 1.11s with PHash)
     public const bool UsingPHash = false; // Enable it to detect duplicates between rotated assets (improve detection) PHash = Perceptual Hash
     // Adjust it if needed, the max advised is less than 90 (for example 68 can detect false positives)
-    public const int PHashThreshold = 40; //By keeping 40, it can detect a Thumbnail and an original with low quality as duplicates; 5 or 6 is often used as a default value in image comparison libraries
+    public const ushort PHashThreshold = 40; //By keeping 40, it can detect a Thumbnail and an original with low quality as duplicates; 5 or 6 is often used as a default value in image comparison libraries
     public const bool UsingDHash = false; // Enable it to Hash in DHash (Difference Hash)
     public const bool DetectThumbnails = false; // Enable it to detect duplicates between a thumbnail and the original
     public const bool AnalyseVideos = false; // Enable it to extract thumbnail from each videos
